Usability improvements to service script
This commit adds support for choosing the compression algorithm used to compress the data stored on the zram swap.

It also simplifies memory size calculation, which is now passed unmodified as KiB to the sysfs parameters of zram devices. This fixed a problem on my Ubuntu 18.04 installation where $MEMORY had a value in scientific notation which was not recognized by the zram module.

Also, it removes unnecesary operations to set the number of zram devices to create.

Finally, it changes the default fraction from 75 to 50, as per recommendation of https://www.kernel.org/doc/Documentation/blockdev/zram.txt, where it states: "There is little point creating a zram of greater than twice the size of memory
since we expect a 2:1 compression ratio. Note that zram uses about 0.1% of the
size of the disk when not in use so a huge zram is wasteful."